requirements.txt
Flask==2.0.2
line_bot_sdk==2.2.1
gunicorn==19.9.0
python-binance
python-dotenv